refactor config code to separate module
This prevents a circular reference in models
